+++ crispy_forms_foundation/layout/buttons.py
@@ -1,5 +1,10 @@
 """
 Button layout items
+
+See :
+
+* `Foundation buttons <http://foundation.zurb.com/docs/components/buttons.html>`_ for button components;
+* `Foundation button groups <http://foundation.zurb.com/docs/components/button_groups.html>`_ for button groups components;
 """
 from django.conf import settings
 from django.template import Context